LiveDateをクラス化
